+++ lending/RELEASES.md
@@ -183,3 +183,13 @@ Based on:
 - [csharp v4.4.0] lending
 ### Releases
 - [NuGet v4.4.0] https://www.nuget.org/packages/Codat.Lending/4.4.0 - lending
+
+## 2023-10-03 00:19:31
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.4.1] lending
+### Releases
+- [NuGet v4.4.1] https://www.nuget.org/packages/Codat.Lending/4.4.1 - lending